Fix time zone offset to be mediumint

+++ sql/mysql/slashschema_create.sql
@@ -603,7 +603,7 @@ CREATE TABLE topics (
 DROP TABLE IF EXISTS tzcodes;
 CREATE TABLE tzcodes (
 	tz char(3) DEFAULT '' NOT NULL,
-	off_set tinyint,
+	off_set mediumint,
 	description varchar(64),
 	PRIMARY KEY (tz)
 );
